(maint) Remove wrap-with-globals
This commit remvoes wrap-with-globals and just passes globals down
through the query-app instead, `assoc`ing the `api-version` at the
top-level so there is no longer a need to have separate files for every
version of the api as most of that logic has been removed from PuppetDB
by now anyways. This change means functions no longer need to rely on
middleware acting on the requests to use data that could be simply
passed through parent functions.