(PDB-1615) Use markers for maplog, not MDC
We were previously using the 'mapped diagnostic context' (MDC) as a way
of passing extra information down to the logger. But the MDC is a
string->string map, so we had to string-encode all values. This makes
them nearly useless for log analysis. 

This change replaces usage of the MDC with the slf4j 'marker' mechanism. 
A marker is an arbitrary object that can be attached to a log message. 
We're using a custom marker that knows how to write json-encoded 
Clojure map data when used in conjunction with
net.logstash.logback.encoder.LoggingEventCompositeJsonEncoder. (This
encoder doesn't require logstash in particular, it just works
well with it)