(PDB-1314) Log sync events with structured data
Start logging sync-related events under the :sync keyword via maplog.
See the coments at the top of sync/command.clj for additional
information.

At the moment, the error handling is preliminary.  The only "special
case" is a 404, which will increment the failed record count and keep
going, but in the future, it sounds like we'll want additional
discrimination.

This commit adds no new tests -- they'll be addressed by (PDB-1613).