(RE-4779) Create a script to migrate config files and any needed data
This commit adds in a script that migrates and updates config files to
aid in the PuppetDB 2 -> PuppetDB 3 upgrade. The script is run in the
postinst during upgrade. It will look either for an archive of config
files created in the preinst, or will look in the PDB 2 location
(/etc/puppetdb/conf.d). If found, it will copy the files over to the new
location (/etc/puppetlabs/puppetdb/conf.c). It will then check the files
to determine if certain settings have been changed from the PDB2 default
setting. If so, that setting is left along. If not, the PDB 2 default
will be updated to reflect PDB 3 standards, and any associated data will
be migrated to the new location. Finally, any old config files are
cleaned up, and a message is printed telling the user to upgrade
puppetdb termini, restart puppetdb, and restart puppetserver.