(PDB-1632) Transfer old AMQP messages at startup
At startup, transfer any messages found in the old
com.puppetlabs.puppetdb.commands queue to puppetlabs.puppetdb.commands
so that pending commands won't be ignored after an upgrade.

Require use of without-jmx instead of always disabling JMX in
with-test-broker, so that we can use with-test-broker in the queue
operation tests.