(maint) Repair broken mq test
This commit fixes a broken mq test where there was an `(= ...)`
expression that wasn't wrapped around its second argument. This failed
silently with a normal `lein test` but caused failures with the
humane-test-output plugin for `lein`. This commit also removes some
unnecessary commas and makes sure to cleanup the `somedir` created
during the mq-tests.