(maint) remove ssl-setup references to AIO pathing
The ssl setup tool uses the existence of the /etc/puppetlabs/puppetdb directory
to assess whether the install target is running PE. This was valid in the olden
days, but with AIO FOSS uses those paths as well. This meant that the ssl
setup tool would set the wrong user in the case of a downgrade from 3.0 to
2.3.x, since at this point the path would exist. This changes the tool to
dispatch on presence of the pe-puppetdb user.