(maint) fix config migration
Without this we were parsing the subname to get the location of the hsqldb data
directory. In the default case though, this resides in the vardir, which is
migrated a moment before.

This patch changes things so that we only attempt to migrate that directory if
the original vardir still exists, which is a signal that the vardir is
nonstandard, but the subname still points to the old one (since it's still the
default value.)