(maint) Fix travis script to checkout similar branch when SNAPSHOT
Currently the test.sh file that Travis uses attempts to find the version
of PuppetDB from the project.clj. If found and version exists as a tag,
it builds that version and uses it. Previously if a tag was not found it
always used master. This commit changes that to use the branch on PDB
that is the same name as the current branch. So if we're on stable, and
no tag is found for "3.0.1-SNAPSHOT" it will checkout stable from PDB
and use that for the test run.