(PDB-1313) Add statistics to puppetdb.commands
Arrange for all command submission and execution to go through the
command-service, so that it can track received and executed command
counts.  This makes it possible to know that all outstanding commands
have been executed by waiting until:

  (let [{:keys [received-commands executed-commands]} (stats cmd-service)]
    (= executed-commands received-commands))

To make that possible, add a PuppetDBCommandDispatcher protocol for
command-service that provides enqueue-command, enqueue-raw-command, and
stats, and make the previous enqueue-command! and enqueue-raw-command!
functions private.  Ensure that all command submissions and executions
update the new stats.