(maint) Add secondsToWaitForCompletion cmd param
When posting a command, you can add a secondsToWaitForCompletion query
parameter. The server will then wait until the command has been
processed before responding to the request, up to the given timeout.

The reponse will include a "processed" key, whose value is JSON true if
the message was processed before the timeout or JSON false if the
timeout was reached. Additionally, if the timeout is reached, the HTTP
response code will be 503.