(maint) Treat the log config as a global resource
Previously pdb-instance invocations would pass a custom config to
trapperkeeper.  Although it probably intended for the changes to be
per-instance, the logging configuration is global, and so whatever
changes were made would apply to everyone, and the original
configuration would never be restored.

As an alternative, add a new suppressing-log-unless-error macro that can
wrap the dynamic lifetime of any number of pdb-instances.  It will
suppress all logging unless there's an event of error level or higher,
in which case it will dump the log to *err* and provide a path to the
temporary log file.

In support of that, add with-additional-log-appenders and
with-log-appenders (and because it was trivial and seems useful,
with-log-event-listener).

Finally, rework the existing operations to use the new ones when
appropriate.

This is unfinished; in particular, the existing pdb-instance invocations
need to be wrapped by suppressing-log-unless-error, and Russell's amqp
suppression filter hasn't been tested yet.