(PDB-1803) Switch report sync to use certname + producer_timestamp
Currently certname + hash is being used to sync reports across PDB
instances. This can cause a potentially nasty upgrade issue if we
change how we compute a hash (something we have done frequently). For
this scenario assume we have two PDB nodes X and Y, X is running an
older version of PDB, Y the newer. Y pulls a new report from X, computes
the hash and persists it. The next time Y pulls from X, it recognizes
that report as a new report (even though it pulled and persisted it)
because it has a different hash. It then submits that report again to
the command queue. It will then recompute the hash and attempt to store
it again which will fail due to a constraint violation (the new hash
already exists in the DB).

These failures would continue on every run, even after PDB X was
ugpraded as we never update report hashes once persisted.