(maint) relocate import/export util functions
Before this patch, we had some functions in our import-export-roundtrip test
that were used in other test namespaces.  This prevented us from using our
fixtures namespace in any of the http tests due to a cyclic dependency, which
I need for PDB-1737. This will break sync testing, but the fix will be
straightforward.