(PDB-1845) Treat the log config as global resource
Previously pdb-instance invocations would pass a custom config to
trapperkeeper.  Although it probably intended for the changes to be
per-instance, the logging configuration is global, and so whatever
changes were made would apply to everyone, and the original
configuration would never be restored.

As an alternative, add a new with-log-suppressed-unless-notable macro
that can wrap the dynamic lifetime of any number of pdb-instances.  It
will suppress all logging unless an event matches its notable-event?
predicate, and if a notable event is encountered, it will dump the log
to *err* and provide a path to the temporary log file.

Include notable-pdb-event?, which ignores the same spurious errors
previously ignored by modifications introduced via the
removed *extra-appender-config*.

To support of the changes, add with-additional-log-appenders and
with-log-appenders (and because it was trivial and seems useful,
with-log-event-listener).

Finally, rework the existing operations to use the new ones when
appropriate.  Rename puppetdb-instance to call-with-puppetdb-instance to
reflect the way it works, and add two new operations:
call-with-lone-log-suppressed-puppetdb-instance and
with-lone-log-suppressed-puppetdb-instance.  They wrap the previous
operations in with-log-suppressed-unless-notable and allow make it
possible to update the previous call sites without additional
wrapping/indentation.