(PDB-1653) Move from get-route to get-registered-endpoints in services.clj
This is short term hack around our startup dependency issues. This was
necessary after the move from many web routing components in PuppetDB to
a single one per top-level context. Previously, the PuppetDBServer
component would ask for it's own route (i.e. (get-route this)). Now this
is handled by the pdb-routing component, but this prefix is needed in
shared-globals for the query components.

The real fix is to break shared-globals out of PuppetDB server and allow
the new pdb-routing namespace to pass it prefix in.