(PDB-1653) PE Extensions changes needed for the PDB top level route
This commit adds a new "pe-routing" service to the extensions
project. With this commit, the top-level (i.e. /pdb) is controlled by
pe-extensions. This allows PE to add additional routes before adding the
handler to the web routing service. This also uses the new maintenance
mode service to ensure 503s are returned while still starting up