(PDB-1852) Add migration for JSONB columns for Unchanged Resources
This commit moves to using JSONB columns for metrics, logs, and the new
column resources in the reports table. The resources column will be used
by unchanged resources for storing "resource" jsonb blobs. The old
metrics and logs json columns have been moved to deprecated_metrics and
deprecated_logs (coalesced with the jsonb columns in the query-eng to
preserve querying for the old data).