(PDB-1852) Store the `resources` doc from the new wire-format of reports
This commit changes PuppetDB's report storage to expect the v6
wire-format with the resources (instead of resource_events) field and
stores the resources document as a JSONB column in reports that was
created in a previous commit's migration addition. This commit updates
the tests to be more compatible with the new version, removing
unnecessary fields in the example reports' `resource_events` fields.