(PDB-1921) Swallow exceptions during the `check-for-update!` on startup
This commit fixes a regression with our check-for-update! function in
which the new shared library we upgraded to does not swallow exceptions
like our previous code. Since the library will log exceptions and
rethrow them, we just need to be sure to swallow the exception in our
wrappers for the library. This commit swallows exceptions for the
version-check library on startup.