(PDB-1732) Remove redundant import/export acceptance tests
This commit remove the import/export acceptance tests as we already have
roundtrip tests in Clojure as well as some acceptance tests to make sure
the bash scripts for the commands themselves work.

This commit also removes redundant anonymization acceptance tests that
were moved to Clojure in a previous commit.

With the remove of the anonymization and import/export acceptance tests,
we can remove the export-comparison helper functions in `acceptance/helpers.rb`.