(PDB-1732) Merge anonymization and export tools
This commit removes the anonymization tool and add the ability for the
export tool to anonymize exports, which allows us to more easily change
the export tool to run with four queries against nodes, reports,
factsets, and catalogs vs. a query for each of those enitities on every
node (on HSQL db we still have to expand each one of those entities).
Because the anonymization tool is now just an export, this commit
adds some of our anonymization acceptance tests in as Clojure tests.