(PDB-1942) Refactor testutils to have better postgresql support
This commit refactors our testutils so that `with-puppetdb-instance`,
used in our more "end-to-end" clojure unit tests, now works with
PostgreSQL. The testutils are also friendlier to use from outside the
FOSS repo, e.g. in `pe-puppetdb-extensions`.

This commit also sets PostgreSQL as the default for unit testing, adding
sane defaults for the database subname/user/password and a script to
create the user/database.