(PDB-1959) Increase JVM permgen space
PDB has been approaching, and recently exceeded the 128M MaxPermSize
setting in project.clj, which results in an out-of-memory exception
during lein test.

The default OpenJDK setting (at least on Linux), appears to be higher
than that (166M), and lein test currently tops out at about 140M.  So
raise the limit to 200M, and don't specify it for anything other than
JDK 7 (since JDK 8 doesn't support it, and issues a warning).