(PDB-1955)(PDB-1698) Run an export in fewer queries
This commit runs an export (without holding all reports/catalogs/facts
in memory as a previous iteration of this change did) in fewer queries
such that we won't make a query for every node's reports/catalogs/facts
but instead just query for all the reports/catalogs/facts at once,
though we still make queries for each entity's child data on hsqldb.
Because we needed to make some refactors to pave the way for making
fewer queries, the code for import/export vs. the cli import/export
components was split out into it's/their own namespace(s). This also
paves the way for easily adding anonymization to the export tool.

Because of the nature of this refactor, we removed some of the export
code which would internally query for a given nodes
reports/catalogs/factsets and sanitize it for resubmission in the import
tool. Therefore some code in our cli tests and in the PE repo had no
equivalent replacement. In order to detangle the export and querying
code, this commit moves some of the useful `get-<entity>` functions from
the PE testutils into the FOSS repo for us to use for testing.

As a side not this commit also adds a facts-only import/export roundtrip
to replace one of our acceptance tests with the same functionality as it
is essentially trivial copy/paste to do so.