(PDB-1943) Actually evaluate anonymizer tests
This commit changes the `for [profile anon/anon-profiles` in the
anonymizer Clojure tests to be `doseq`s instead so that they actually
get run.

This commit also fixes an issue where the export tool version of the
anoymizer was not passing `anonymize-fn` correctly due to a type. This
shed light on an issue where the anonymization tool only works on
entities with string keys, so this commit also makes sure we stringify
and then keywordize on the way out for anonymizing our entities on
export.