(PDB-1984) Migrate benchmarking tool to submit v6 reports
This commit migrates the benchmarking tool to "sanitize" its example v5
wireformat reports to v6 reports, and removes some unnecessary json
generating and unparsing we do in the tool.
Since we're removing the json generating and parsing for reports, the
same unnecessary json roundtrips are removed from catalogs and facts as
well for consistency.