(PDB-1810) Add code-id field to catalogs
This commit adds the code-id field to catalogs, adding a migrations to
add the column to the catalogs schema, adding storage of the field, and
adding querying of the field to PuppetDB. This commit also updates the
terminus to add the field to catalogs on submission. This commit
consolidates some of the wireformat logic in catalogs.clj, aligning it
more with what we have in PuppetDB on the reports side of things.
We expect code-id will be null for FOSS PuppetDB for the forseable
future.