(PDB-2039) WIP - Proof of concept code on securing at the node level
This commit hacks in the ssl client cert to the JMS message header to be
validated when the command is pulled off the queue.

There's still plenty of work left. Currently this is just a hack. There
needs to be a config option to turn this on. It needs to be added to
each command type (currently it's only on catalogs) and we also need to
have a cert whitelist check. There's probably a bunch of broken tests too.