(pdb-2034) schema error on replace-facts with escape-quoted strings
Without this commit we're broken on escape-quoted fact names and fact names
containing #~. Some escape-quoting we do before storage was causing different
names to come out of realize-paths! than went in, which caused an empty key
lookup that led to the error.

This commit stops stripping escaped quotes in fact-contents result munging in
all cases except escape-quoted numbers. In the numbers case it's necessary
because we encode fact paths as strings, but still need to be able to
distinguish between actual numbers (stored quoted) and stringified numbers
(stored escape-quoted).