(maint) Add `:exclusions [org.slf4j/slf4j-api]` for structured-logging
This commit adds the slf4j-api exclusion for the structured-logging
library because our packaging jobs for pe-puppetdb are picking up a
conflict with activemq and the structured-logging dependencies here
(with the slf4j-api version). This does not appear to show up with a
lein install/test but does show up in the packaging job.