(PDB-2096) Add not-found route to top-level `/pdb` route
This commit adds the `compojure.route/not-found` route at the top-level
of the PuppetDB routes so that new PuppetDB routes won't need to add the
not-found route. This also fixes an error where if a user hit any
non-existent route under `/pdb`, PDB would throw an exception related to
our middleware not playing nicely with not-founds which produced a
confusing `nil...status` error.