(maint) Stop encoding json child data twice
This commit removes the munge functions for catalogs, reports and
factsets that were parsing the json child data for those entities. This
commit adds a json encoder that can handle encoding the raw database
json objects so we don't encode the child data twice.

This commit also adds optional pretty-printing to PuppetDB which
requires us to parse and re-encode JSON data from the database.
Note that the default is to not pretty-print the JSON.