(PDB-2122) New 'root' single query endpoint and 'from' operator
This patch introduces the new 'root' endpoint that allows for querying any
entity inside a single query by providing a context query.

Here I've included this new feature all the way back to produce-streaming-body,
stream-query-result and the query API. This patch means that all the HTTP
endpoints are required to wrap their queries in from, so that this concept
of a single query for all endpoints exists all the way back to the query engine
API.

For cleanup, I've modified the produce-streaming-body and stream-query-results
to take an options map instead of a list of ordered arguments as we seem to change
this frequently and some validation is desired. Plus this simplifies argument
passing as most of the data is already available in `globals`.

As a side-effect of adding the from operator, I've added a new format for
subqueries that uses the same syntax at this level, as it does on the top-level.
This is provided for consistency, and it will hopefully remove the need for
the select_<entity> operator which was always a bit special.

Also, documentation for entities is now centralised.

Signed-off-by: Ken Barber <ken@bob.sh>