(MAINT) Don't fail when old migration versions are present
This fixes a bug created by the collapsed 1.x and 2.x migrations. The
bug would occur in upgrade scenarios. If a pre-3.0 migration version was
found in the database, it would fail, even if the database was a 3.0.0
database (and upgrades are supported). This commit fixes that and
separates out the supported upgrade versions and the known previous
version that are no longer referenced in PuppetDB.