(PDB-2097) Disable Puppet Server startup checkins to Dujour
Prior to this commit Puppet Server would checkin to Dujour when
puppetserver was started up because we munged `/etc/hosts` after
puppetserver had already been installed. This commit separates the
installation of puppet and puppetserver so that we can put the munging
of `/etc/hosts` in the middle.

This commit also removes a lot of dead code that was getting in the way
of the refactor.