(PDB-2075) Clean up with-puppetdb-instance test dbs
Instead of creating a database in create-temp-config, handle the
remaining temporary databases via with-test-db (including calls to
with-puppetdb-instance).

In order to deal with the fact that at least our version of
BoneCP (0.7.1) doesn't appear to properly clean up its DB connections on
close[1], forcibly eject the test user before dropping the test databases.

[1] Using BoneCP 0.8.0 appears to fix the problem, but we can't upgrade
    because it incompatibly changed its error handling.