(MAINT) Switch to connecting to the postgres table for tests
The current code connects to the template1 database to setup the
database test templates. This doesn't work in a shared environment
because Postgres won't allow you to copy a database when there is a user
connected to the database. In our case we were connecting to template1,
and template1 was being used to create the template database that all of
our tests would use.

This commit instead connects to the postgres database, which should
always be present and should not cause issues in creating new databases
derived from template1