(PDB-1927) PQL translation to AST
This patch introduces the new Puppet query language to PuppetDB. It provides a basic
parse for this capability, documentation for the new language and a hook in the
index endpoint for the HTTP API side.

As this work is fairly new and extensive, some of the expected work has been
broken into future tickets. For now, this is marked as experimental until we are
happy with the language.

Signed-off-by: Ken Barber <ken@bob.sh>