(BKR-643) Introduce and use Beaker::DSL.register
This is a spike commit to see if we can get this part of the acceptance suite
green using the most recent Beaker. Changes to support BKR-623, introduced
via https://github.com/puppetlabs/beaker/pull/1013 result in relocation of
`Beaker::TestCase`-specific code (as the `Beaker::TestSuite` and
`Beaker::TestCase` modules are test-runner specific, and the nature of
BKR-623 is to lay the groundwork to support multiple test runners).

Test suites which are overly familiar with the `Beaker::Test*` hierarchy are
seeing scoping issues, where injection of methods into the `Beaker::TestCase`
module ends up not providing the right scope.

There are a few ways to go about fixing this; this commit introduces one. The
registration method was already developed for a later stage of the test runner
refactorings (see: https://github.com/puppetlabs/beaker/pull/1023). We wish to
gradually remove uses of `Beaker::TestCase` from non-Beaker code, and exposing
a regisration facility on `Beaker::DSL` is our preferred method.

Backporting this registration facility here to see if it allows us to go green.