(PDB-2234) Upgrade our metrics-clojure dependency from `0.7.0`->`2.6.0`
This commit updates our metrics-clojure dependency from `0.7.0` to
`2.6.0` which uses the dropwizard java metrics library. This update
means that we register our metrics with a metrics registry, which is the
object we will need in order to register HikariCP metrics. This commit
does just enough to upgrade and does not refactor our metrics code.
Notable future changes that should be done in a follow-up patch, would
be to register all our metrics on start up, and have functions which use
those metrics, fetch the metric object from the registry rather than
storing the metric object in the atoms we have today as well as the
registries.

This commits also adds an explicit registering of the MetricsRegistry
with the JMXReporter from Dropwizard metrics, required by their API in
the newest version. We start the reporting during the trapperkeeper init
phase and stop the reporting during the trapperkeeper stop phase.

This commit also cleans up the `puppetlabs.puppetdb.metrics.core`
namespace, to separate concerns with the http handlers which use the
namespace in `puppetlabs.puppetdb.metrics`. This allows us to keep the
`metrics-registry` and `jmx-metrics-reporter` objects in a sane
location, i.e. no cyclic dependencies between the middleware and metrics
namespaces.