(PDB-2239) Use encode for handling bytea data
This commit changes our handling of bytea from using `trim` (which
relies implicitly on `bytea_output` being set to `'hex'` in PostgreSQL)
to instead use `encode(...,'hex')`. This allows our `bytea` usage to be
database setting independent.

This commit also cleans up some of the `fact_values` code in storage.clj
which was using `parse-db-hash` which is also not `bytea_output`
independent.