(maint) Remove paging-option parsing duplication
This commit removes the `wrap-with-paging-options` middleware from
PuppetDB which was duplicating code from the `http/query.clj` namespace
and whose only function was to catch `IllegalArgumentException`s and
return an `error-response`. This commit adds this exception catching to
it's own `wrap-with-illegal-arugment-catch` middleware. This commit also
moves the illegal-argument-catch and the verify-accepts-json middleware
to wrap the whole of the Query-API app which greatly simplifies the code
in each of the endpoints.