(PDB-2286) Add ability to always store catalog jsonb fields
This commit fixes a bug with historical catalogs where a user could turn
off historical catalogs and break some of the PE views because the JSONB
column storage was also turned off. This new atom will always be set to
true in PE (as opposed to the historical-catalogs' atom which is config
dependent).