(PDB-2286) Make the merge between resources more robust
This commit makes the resources merge between catalogs and reports more
robust by making sure that reports or catalogs having a nil entry for
resources won't cause the response to be nil, which was happening
erroneously because of the behavior of the `clojure.set/join` function.

(PDB-2288) Alter historical-catalogs response format for resources

This commit alters the `historical-catalogs` response format to align it
with `catalogs` response format, for export and sync code.