(PDB-2252) Try to avoid parsing a POSTed cmd
Don't parse the incoming command body when it has command and version
parameters (which should now be the common case).  Instead, validate the
parameters, and pass the body directly to the queue, uninterpreted.
Otherwise, warn if body has to be parsed into RAM.

Attach the command, version, and if available, certname, to the queue
message as JMS properties, extracting them from the body (after
parsing), if they weren't available in the incoming request.

As before, reject incoming commands that don't have both command and
version when they have either.