(maint) use table stats to compute the number of resources
This commit changes one of our population metrics to use table stats to compute
the number of resources. It also eliminates a restriction to active nodes for
the metric, and splits the num-nodes metric into num-active-nodes and
num-inactive-nodes in the dashboard to make things clearer. num-nodes is left
around for backward compatibility. Lastly it adds a mechanism for independent
specification of chart options for our metrics, and changes the
percent-resource-duplication query to run every 30 seconds instead of 5. This
may or may not be the right number.