(PDB-2214) anonymize words as ?????
This fixes an anonymization bug where we would anonymize text blobs as
completely different, sometimes long random strings even if they only differed
by a word. This prevented any reasonable compression of report logs/messages,
which meant that the tool produced unmanagebly large tarballs on big installs.

With this patch a text blob is replaced by a string of question marks of the
same length.