(PDB-2380) Limit the storage of historical-catalogs
This commit limits the storage of historical catalogs to only the latest
three catalogs for a node. This commit limits the storage because for
our benchmarking data with 5000 nodes (~3 million catalogs) we were
seeing the catalogs table take up around 140GB of space.

This commit also changes the `store-catalogs-historically?` setting for
PE to be a limit of the number of previous historical-catalogs to store.