(maint) Stop requiring command_versions in import requests
This commit changes import to no longer require command_versions in the
import request and instead assumes that we can parse metadata.json from
the first entry of the archive tarball (or rather we discard every entry
in the tarball until we find the metadata). This will allow our CLI
tooling to not have to parse the JSON from the tarball.