(PDB-2424) Fix delete-edges! to hit unique index
Currently delete-edges coerces the column and compares it with the data
passed in. This cause ALL source/target column values to be
decoded. This is slow and doesn't hit the unique index that is in
place. Instead this patch converts the source/target being deleted to
bytea and compares it with what is stored. This hits the index and will
be much faster as only 1 conversion is done, not one for each row + column.