(PDB-2737) Split message consumers from processing
This commit separates the thread listening to ActiveMQ and the threads
that are processing the message (and persisting the changes to
disk). The intent of this change is to factor the code in a way that
allows a new queuing system to be dropped in place of ActiveMQ. The code
as it is has a subtle but relating to how JMS expects messages to be
acknowledged. When acknowledging a message, it acknowledges all
previously unacknowledged messages for that thread. This code consumes all
messages (from ActiveMQ's perspective) via a single thread. This means it's
possible for a slow message to be in-flight, and a faster/smaller
message that is processed after, to acknolwedge that slower message. If
that slower message was then to fail, we'd have problems as ActiveMQ
would already have marked the message as successfully processed.

Once we have removed ActiveMQ from this pipeline, we'll be able to
acknowledge the message(s) in a way that corresponds with our usage.