(maint) Fix flaky reports test
Now that we do the post-migration vacuum in the background, it can
race with test execution (especially in the 3.x series, where we are
still running db migration once per test case). In this case, we are
testing pagination. If we do the first paginated query, then vacuum
executes, and then we do the second paginated query, the underlying
table can be rewritten to have a different natural order.

Fix this by supplying an order_by to the paginated query, ensuring
its stability across storage change.