(maint) Remove some potential ordering issues from tests
With the inclusion of the recent backgrounded 'analyze' it introduced some
flakiness to our tests due to ordering issues. This patch fixes two cases
where we are doing a limit without an order-by, and converts resource_events
for reports tests to a set so comparison always works.

This patch also fixed munge-tar-map, so that its munging on map values instead
of its prior expectation. Beforehand this function was simply creating new keys
in the top-evel map, instead of operating on existing keys inside the values.

Signed-off-by: Ken Barber <ken@bob.sh>